pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-agents"
version = "0.1.0"
description = "Godot integration helpers for llama.cpp"
authors = [{name = "Local Agents"}]
requires-python = ">=3.10"
dependencies = [
  "huggingface-hub>=0.23",
  "llama-cpp-python>=0.2.90",
  "pytest"
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools]
packages = ["local_agents"]

[tool.pytest.ini_options]
addopts = "-ra"
